Upgrade from Chromium 105 to Chromium 106 (1.44.x)

+++ browser/ui/views/sidebar/sidebar_control_view.h
@@ -72,9 +72,9 @@ class SidebarControlView : public views::View,
 
   // sidebar::SidebarModel::Observer overrides:
   void OnItemAdded(const sidebar::SidebarItem& item,
-                   int index,
+                   size_t index,
                    bool user_gesture) override;
-  void OnItemRemoved(int index) override;
+  void OnItemRemoved(size_t index) override;
 
   void Update();
